rename parts folder to components
