+++ src/pages/main/components/settings.js
